backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic[email]
